DataPlatform/requirements.txt
Django==4.1.3
django-allauth==0.51.0
psycopg2-binary>=2.9
kafka-python
confluent-kafka
google-api-python-client
google-auth-httplib2
google-auth-oauthlib